Updated on 28th June 2020. The index.html page is responsive on mobile and tablet. However, some images are not responsive on tablet and need to be carefully looked at. In the meantime, I have started creating the second page, which focuses on the medical UX project. On this page, I have already finished the first two sections and now have to progress on the other sections.
Updated on 28th June 2020. The index.html page is responsive on mobile and tablet. However, some images are not responsive on tablet and need to be carefully looked at. In the meantime, I have started creating the second page, which focuses on the medical UX project. On this page, I have already finished the first two sections and now have to progress on the other sections.