Updated on 28th June 2020. The index.html page is now responsive across all devices (mobile, tablet, and laptop). I encountered numerous issues with the responsiveness and now I have resolved them with the help of Bootstrap. Now, I need to focus on the other page and complete that on time.
Updated on 28th June 2020. The index.html page is now responsive across all devices (mobile, tablet, and laptop). I encountered numerous issues with the responsiveness and now I have resolved them with the help of Bootstrap. Now, I need to focus on the other page and complete that on time.